Removing the data and code pertaining to treestore.
Removing the data and code that pertained only to the original, and long defunct
use of this repo for an experimental treestore. You can get back that state of
the repository through tag legacy-treestore. See [phylotastic/rdf-treestore](http://github.com/phylotastic/rdf-treestore) for a reincarnation and actively developed version of an RDF treestore.